结构拆分
